+++ src/components/LaborLecture/CallToAction.js
